requirements.txt
Django==4.2.5
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
cryptography==41.0.4
defusedxml==0.7.1
dj-database-url==2.1.0
dj-rest-auth==5.0.1
django-allauth==0.57.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
idna==2.8
oauthlib==3.2.2
packaging==23.2
psycopg2-binary==2.9.9
pycparser==2.21
pyjwt==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.0
requests==2.22.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing-extensions==4.8.0
urllib3==1.25.11